requirements.txt
auto_mix_prep==0.2.0
backpack==0.1
hydra-core==1.3.1
matplotlib==3.6.1
mlxtend==0.21.0
numpy==1.23.4
omegaconf==2.2.3
pandas==1.5.1
Pillow==9.4.0
pyrootutils==1.0.4
pytorch_lightning==1.7.1
rich==13.0.0
torch==1.13.0
torchmetrics==0.9.3
torchvision==0.14.0
tqdm==4.62.2
wandb==0.13.4
wilds==2.0.0
